Rename window.h to lutro_window.h to match the source file
